load from source asset instead of FromRootInput when working with assets
