+++ python_modules/dagster/dagster/_core/execution/plan/plan.py
@@ -499,13 +499,13 @@ def get_step_input_source(
         #  make sure input is unconnected in the outer dependency structure too
         not solid.container_maps_input(input_handle.input_name)
     ):
-        if input_def.root_manager_key or input_def.input_manager_key:
-            return FromRootInputManager(solid_handle=handle, input_name=input_name)
         # can only load from source asset if assets defs are available
-        elif asset_layer.has_assets_defs and asset_layer.asset_key_for_input(
+        if asset_layer.has_assets_defs and asset_layer.asset_key_for_input(
             handle, input_handle.input_name
         ):
             return FromSourceAsset(solid_handle=handle, input_name=input_name)
+        elif input_def.root_manager_key or input_def.input_manager_key:
+            return FromRootInputManager(solid_handle=handle, input_name=input_name)
 
     if dependency_structure.has_direct_dep(input_handle):
         solid_output_handle = dependency_structure.get_direct_dep(input_handle)
